requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.6
djangorestframework==3.15.2
pillow==11.1.0
reportlab==4.3.1
sqlparse==0.5.3
tzdata==2025.1
